requirements.txt
scikit-learn>=0.24.1
spacy>=2.0.18
python-docx>=0.8.10
Flask>=1.1.1
Flask-RESTful>=0.3.8
pandas>=1.2.3
numpy>=1.19.5